Add Fast Refresh
Moved constants out of file with React components.
